chore: remove product page component
